requirements.txt
langchain
langchain-huggingface
streamlit
transformers
streamlit-chat

faiss-cpu
boto3
langchain-community
python-dotenv
langchain-openai
pypdf
huggingface_hub
accelerate
bitsandbytes
wikipedia
numexpr